requirements.txt
Flask
gunicorn
numpy
Pillow
tqdm
faiss-cpu
sentence-transformers
torch
torchvision